requirements.txt
py-cord @ git+https://github.com/Pycord-Development/pycord
ezcord
python-dotenv
pygount